refactor: 합/불 낙관적 업데이트 공통 훅 처리
